[None][fix] Unblock CI on PR 13340 — yapf format + stubgen torch preload
- tensorrt_llm/_torch/modules/attention.py: apply yapf's canonical
  continuation style to the new_empty() calls inside
  _mla_dsa_proj_fake's FP4 branch. The round-3 commit used the
  arg-on-next-line layout, which the Release-Check `yapf` hook rejects.

- scripts/build_wheel.py: pre-import torch before invoking
  pybind11_stubgen for deep_gemm_cpp_tllm. The DeepGEMM c491439e bump
  registers FP4 / UE8M0 scalar types into torch at module init, so the
  pybind11_stubgen subprocess fails with `ImportError: unsupported
  scalarType` because it loads the module via a plain
  importlib.import_module without touching torch. Wrapping the call in
  `python -c 'import torch; from pybind11_stubgen import main; main()'`
  reuses the existing CLI args (argparse reads sys.argv[1:]) so the
  rest of the invocation is unchanged.

Verified:
  - yapf --diff tensorrt_llm/_torch/modules/attention.py → no diff.
  - Full `python scripts/build_wheel.py --install` (no --skip-stubs)
    succeeds and produces tensorrt_llm/deep_gemm_cpp_tllm.pyi.

+++ tensorrt_llm/_torch/modules/attention.py
@@ -1060,11 +1060,11 @@ def _mla_dsa_proj_fake(
         q_fp8 = hidden_states.new_empty(
             [num_tokens, indexer.n_heads, indexer.head_dim // 2],
             dtype=torch.int8)
-        k_fp8 = hidden_states.new_empty(
-            [num_tokens, indexer.head_dim // 2], dtype=torch.int8)
+        k_fp8 = hidden_states.new_empty([num_tokens, indexer.head_dim // 2],
+                                        dtype=torch.int8)
         k_scale = hidden_states.new_empty([num_tokens, 1], dtype=torch.int32)
-        q_scale = hidden_states.new_empty(
-            [num_tokens, indexer.n_heads, 1], dtype=torch.int32)
+        q_scale = hidden_states.new_empty([num_tokens, indexer.n_heads, 1],
+                                          dtype=torch.int32)
     else:
         q_fp8 = hidden_states.new_empty(
             [num_tokens, indexer.n_heads, indexer.head_dim],
